# backend/scratch/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AIHire Recruitment Platform",
    "description": "Backend API service for the AI-powered Recruitment Platform.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/otp/send": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Send Otp",
        "operationId": "send_otp_auth_otp_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOTPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/otp/verify": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Verify Otp Api",
        "operationId": "verify_otp_api_auth_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOTPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/otp/last": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get Last Otp Api",
        "description": "Returns the last generated OTP. Useful for testing and development.",
        "operationId": "get_last_otp_api_auth_otp_last_get",
        "parameters": [
          {
            "name": "target",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Target"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get Current User",
        "operationId": "get_current_user_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/candidate/resume/upload": {
      "post": {
        "tags": [
          "Candidate Resumes"
        ],
        "summary": "Upload Candidate Resume",
        "operationId": "upload_candidate_resume_candidate_resume_upload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateResumeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateResumeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/candidate/resume/{user_id}": {
      "get": {
        "tags": [
          "Candidate Resumes"
        ],
        "summary": "List Candidate Resumes",
        "operationId": "list_candidate_resumes_candidate_resume__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateResumeResponse"
                  },
                  "title": "Response List Candidate Resumes Candidate Resume  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/resume-parser/upload": {
      "post": {
        "tags": [
          "Resume Parser"
        ],
        "summary": "Upload and parse a resume, storing results in the database",
        "description": "Full pipeline:\n    Upload \u2192 Local Storage \u2192 Text Extraction \u2192 Cleaning\n    \u2192 Parsing \u2192 JSON Generation \u2192 Store in resume_parser_results",
        "operationId": "upload_resume_resume_parser_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_resume_parser_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeParserResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/resume-parser/parse": {
      "post": {
        "tags": [
          "Resume Parser"
        ],
        "summary": "Parse a resume and return structured JSON without storing",
        "description": "Stateless parse \u2014 no database record created.\nReturns raw_text, cleaned_text, and parsed_json.",
        "operationId": "parse_resume_resume_parser_parse_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_resume_resume_parser_parse_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseOnlyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/resume-parser/": {
      "get": {
        "tags": [
          "Resume Parser"
        ],
        "summary": "List all parsed resumes for the authenticated user",
        "description": "Returns a summary list (without raw/cleaned text) for the authenticated user.",
        "operationId": "list_resumes_resume_parser__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "parsing_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status: pending | processing | completed | failed",
              "title": "Parsing Status"
            },
            "description": "Filter by status: pending | processing | completed | failed"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset",
              "default": 0,
              "title": "Skip"
            },
            "description": "Pagination offset"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Pagination limit",
              "default": 50,
              "title": "Limit"
            },
            "description": "Pagination limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumeParserResultSummary"
                  },
                  "title": "Response List Resumes Resume Parser  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/resume-parser/{record_id}": {
      "get": {
        "tags": [
          "Resume Parser"
        ],
        "summary": "Get a full resume parser result by ID",
        "description": "Returns the full record including raw_text, cleaned_text, and parsed_json.",
        "operationId": "get_resume_resume_parser__record_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeParserResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Resume Parser"
        ],
        "summary": "Delete a resume parser result by ID",
        "description": "Deletes the record. Returns 404 if not found, 403 if not owned by caller.",
        "operationId": "delete_resume_resume_parser__record_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_parse_resume_resume_parser_parse_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Resume file (PDF, DOCX, DOC, TXT, JPG, PNG)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_parse_resume_resume_parser_parse_post"
      },
...
      "Body_upload_resume_resume_parser_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Resume file (PDF, DOCX, DOC, TXT, JPG, PNG)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_resume_resume_parser_upload_post"
      },
...
      "CandidateResumeCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "resume_name": {
            "type": "string",
            "maxLength": 255,
            "title": "Resume Name"
          },
          "resume_path": {
            "type": "string",
            "maxLength": 1000,
            "title": "Resume Path"
          },
          "file_type": {
            "type": "string",
            "maxLength": 50,
            "title": "File Type"
          },
          "file_size": {
            "type": "integer",
            "minimum": 0.0,
            "title": "File Size"
          },
          "ats_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Ats Score"
          },
          "parsed_status": {
            "type": "boolean",
            "title": "Parsed Status",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "resume_name",
          "resume_path",
          "file_type",
          "file_size"
        ],
        "title": "CandidateResumeCreate"
      },
      "CandidateResumeResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "resume_name": {
            "type": "string",
            "title": "Resume Name"
          },
          "resume_path": {
            "type": "string",
            "title": "Resume Path"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "ats_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ats Score"
          },
          "parsed_status": {
            "type": "boolean",
            "title": "Parsed Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "resume_name",
          "resume_path",
          "file_type",
          "file_size",
          "ats_score",
          "parsed_status",
          "created_at",
          "updated_at"
        ],
        "title": "CandidateResumeResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ParseOnlyResponse": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "raw_text": {
            "type": "string",
            "title": "Raw Text"
          },
          "cleaned_text": {
            "type": "string",
            "title": "Cleaned Text"
          },
          "parsed_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parsed Json"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "raw_text",
          "cleaned_text",
          "parsed_json"
        ],
        "title": "ParseOnlyResponse",
        "description": "Response for parse-only endpoint (no database record created)."
      },
...
      "ResumeParserResultResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "resume_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume File"
          },
          "file_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Type"
          },
          "raw_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Text"
          },
          "cleaned_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cleaned Text"
          },
          "parsed_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed Json"
          },
          "parsing_status": {
            "type": "string",
            "title": "Parsing Status"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "candidate_id",
          "original_filename",
          "resume_file",
          "file_type",
          "raw_text",
          "cleaned_text",
          "parsed_json",
          "parsing_status",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "ResumeParserResultResponse",
        "description": "Full response schema for a resume_parser_results record."
      },
      "ResumeParserResultSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "resume_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume File"
          },
          "file_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Type"
          },
          "parsing_status": {
            "type": "string",
            "title": "Parsing Status"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "candidate_id",
          "original_filename",
          "resume_file",
          "file_type",
          "parsing_status",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "ResumeParserResultSummary",
        "description": "Lightweight list-view response (excludes raw/cleaned text)."
      },
...
      "SendOTPRequest": {
        "properties": {
          "method": {
            "type": "string",
            "title": "Method",
            "description": "email or phone"
          },
          "target": {
            "type": "string",
            "title": "Target"
          }
        },
        "type": "object",
        "required": [
          "method",
          "target"
        ],
        "title": "SendOTPRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyOTPRequest": {
        "properties": {
          "method": {
            "type": "string",
            "title": "Method"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "code": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "title": "Code"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "method",
          "target",
          "code",
          "role"
        ],
        "title": "VerifyOTPRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}